Samsung Phone Advisor/general/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
logging
time
json
re
urllib.parse
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
